Solve circular dependency by creating a separate workflow creation service
