+++ packages/cli/src/workflows/__tests__/workflow.service.test.ts
@@ -1,18 +1,14 @@
-import type { LicenseState } from '@n8n/backend-common';
-import type { User, ProjectRepository } from '@n8n/db';
-import { WorkflowEntity } from '@n8n/db';
+import type { User } from '@n8n/db';
+import type { WorkflowEntity } from '@n8n/db';
 import type { Scope } from '@n8n/permissions';
 import type { MockProxy } from 'jest-mock-extended';
 import { mock } from 'jest-mock-extended';
 
-import type { CredentialsService } from '@/credentials/credentials.service';
 import { userHasScopes } from '@/permissions.ee/check-access';
-import type { ProjectService } from '@/services/project.service.ee';
 import type { RoleService } from '@/services/role.service';
 import type { WebhookService } from '@/webhooks/webhook.service';
 import type { WorkflowFinderService } from '@/workflows/workflow-finder.service';
 import { WorkflowService } from '@/workflows/workflow.service';
-import type { EnterpriseWorkflowService } from '@/workflows/workflow.service.ee';
 import * as WorkflowHelpers from '@/workflow-helpers';
 
 jest.mock('@/permissions.ee/check-access');
@@ -64,10 +60,6 @@ describe('WorkflowService', () => {
 				webhookServiceMock, // webhookService
 				mock(), // licenseState
 				mock(), // projectRepository
-				mock(), // tagRepository
-				mock(), // credentialsService
-				mock(), // folderService
-				mock(), // enterpriseWorkflowService
 			);
 		});
 
@@ -205,10 +197,6 @@ describe('WorkflowService', () => {
 				mock(), // webhookService
 				mock(), // licenseState
 				mock(), // projectRepository
-				mock(), // tagRepository
-				mock(), // credentialsService
-				mock(), // folderService
-				mock(), // enterpriseWorkflowService
 			);
 
 			jest.clearAllMocks();
@@ -312,263 +300,6 @@ describe('WorkflowService', () => {
 			});
 
 			expect(userHasScopesMock).not.toHaveBeenCalled();
-		});
-	});
-
-	describe('createWorkflow()', () => {
-		describe('credential retrieval', () => {
-			test('should include global credentials when checking credential permissions', async () => {
-				const credentialsServiceMock = mock<CredentialsService>();
-				const enterpriseWorkflowServiceMock = mock<EnterpriseWorkflowService>();
-				const licenseStateMock = mock<LicenseState>();
-				const projectServiceMock = mock<ProjectService>();
-				const projectRepositoryMock = mock<ProjectRepository>();
-
-				credentialsServiceMock.getMany.mockResolvedValue([]);
-				licenseStateMock.isSharingLicensed.mockReturnValue(true);
-				enterpriseWorkflowServiceMock.validateCredentialPermissionsToUser.mockImplementation(() => {
-					throw new Error('Stopping for test');
-				});
-				projectServiceMock.getProjectWithScope.mockResolvedValue({ id: 'project-1' } as never);
-
-				const workflowService = new WorkflowService(
-					mock(), // logger
-					mock(), // sharedWorkflowRepository
-					mock(), // workflowRepository
-					mock(), // workflowTagMappingRepository
-					mock(), // binaryDataService
-					mock(), // ownershipService
-					mock(), // tagService
-					mock(), // workflowHistoryService
-					mock(), // externalHooks
-					mock(), // activeWorkflowManager
-					mock(), // roleService
-					projectServiceMock, // projectService
-					mock(), // executionRepository
-					mock(), // eventService
-					mock(), // globalConfig
-					mock(), // folderRepository
-					mock(), // workflowFinderService
-					mock(), // workflowPublishHistoryRepository
-					mock(), // workflowValidationService
-					mock(), // nodeTypes
-					mock(), // webhookService
-					licenseStateMock, // licenseState
-					projectRepositoryMock, // projectRepository
-					mock(), // tagRepository
-					credentialsServiceMock, // credentialsService
-					mock(), // folderService
-					enterpriseWorkflowServiceMock, // enterpriseWorkflowService
-				);
-
-				const user = mock<User>();
-				const newWorkflow = new WorkflowEntity();
-				newWorkflow.name = 'Test';
-				newWorkflow.nodes = [];
-				newWorkflow.connections = {};
-
-				await expect(
-					workflowService.createWorkflow(user, newWorkflow, { projectId: 'project-1' }),
-				).rejects.toThrow();
-
-				expect(credentialsServiceMock.getMany).toHaveBeenCalledWith(user, {
-					includeGlobal: true,
-				});
-			});
-		});
-
-		test('should throw BadRequestError when user lacks access to credentials in workflow', async () => {
-			const credentialsServiceMock = mock<CredentialsService>();
-			const enterpriseWorkflowServiceMock = mock<EnterpriseWorkflowService>();
-			const licenseStateMock = mock<LicenseState>();
-			const projectServiceMock = mock<ProjectService>();
-			const projectRepositoryMock = mock<ProjectRepository>();
-
-			credentialsServiceMock.getMany.mockResolvedValue([]);
-			licenseStateMock.isSharingLicensed.mockReturnValue(true);
-			enterpriseWorkflowServiceMock.validateCredentialPermissionsToUser.mockImplementation(() => {
-				throw new Error('User does not have access');
-			});
-			projectServiceMock.getProjectWithScope.mockResolvedValue({ id: 'project-1' } as never);
-
-			const workflowService = new WorkflowService(
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				projectServiceMock,
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				licenseStateMock,
-				projectRepositoryMock,
-				mock(),
-				credentialsServiceMock,
-				mock(),
-				enterpriseWorkflowServiceMock,
-			);
-
-			const user = mock<User>();
-			const newWorkflow = new WorkflowEntity();
-			newWorkflow.name = 'Test';
-			newWorkflow.nodes = [];
-			newWorkflow.connections = {};
-
-			await expect(
-				workflowService.createWorkflow(user, newWorkflow, { projectId: 'project-1' }),
-			).rejects.toThrow(
-				'The workflow you are trying to save contains credentials that are not shared with you',
-			);
-		});
-	});
-
-	describe('redaction policy scope enforcement on create', () => {
-		const userHasScopesMock = jest.mocked(userHasScopes);
-
-		test('should strip redactionPolicy when user lacks scope', async () => {
-			const projectServiceMock = mock<ProjectService>();
-			const projectRepositoryMock = mock<ProjectRepository>();
-			const licenseStateMock = mock<LicenseState>();
-
-			projectServiceMock.getProjectWithScope.mockResolvedValue({ id: 'project-1' } as never);
-			licenseStateMock.isSharingLicensed.mockReturnValue(false);
-			userHasScopesMock.mockResolvedValue(false);
-
-			const transactionManager = {
-				save: jest.fn().mockRejectedValue(new Error('Stopping for test')),
-			};
-			(projectRepositoryMock as any).manager = {
-				transaction: jest.fn(async (cb: any) => cb(transactionManager)),
-			};
-
-			const workflowService = new WorkflowService(
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				projectServiceMock,
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				licenseStateMock,
-				projectRepositoryMock,
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-			);
-
-			const user = mock<User>();
-			const newWorkflow = new WorkflowEntity();
-			newWorkflow.name = 'Test';
-			newWorkflow.nodes = [];
-			newWorkflow.connections = {};
-			newWorkflow.settings = { redactionPolicy: 'all' };
-
-			await expect(
-				workflowService.createWorkflow(user, newWorkflow, { projectId: 'project-1' }),
-			).rejects.toThrow('Stopping for test');
-
-			expect(userHasScopesMock).toHaveBeenCalledWith(
-				user,
-				['workflow:updateRedactionSetting'],
-				false,
-				{ projectId: 'project-1' },
-			);
-		});
-
-		test('should preserve redactionPolicy when user has scope', async () => {
-			const projectServiceMock = mock<ProjectService>();
-			const projectRepositoryMock = mock<ProjectRepository>();
-			const licenseStateMock = mock<LicenseState>();
-
-			projectServiceMock.getProjectWithScope.mockResolvedValue({ id: 'project-1' } as never);
-			licenseStateMock.isSharingLicensed.mockReturnValue(false);
-			userHasScopesMock.mockResolvedValue(true);
-
-			const transactionManager = {
-				save: jest.fn().mockRejectedValue(new Error('Stopping for test')),
-			};
-			(projectRepositoryMock as any).manager = {
-				transaction: jest.fn(async (cb: any) => cb(transactionManager)),
-			};
-
-			const workflowService = new WorkflowService(
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				projectServiceMock,
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-				licenseStateMock,
-				projectRepositoryMock,
-				mock(),
-				mock(),
-				mock(),
-				mock(),
-			);
-
-			const user = mock<User>();
-			const newWorkflow = new WorkflowEntity();
-			newWorkflow.name = 'Test';
-			newWorkflow.nodes = [];
-			newWorkflow.connections = {};
-			newWorkflow.settings = { redactionPolicy: 'all' };
-
-			await expect(
-				workflowService.createWorkflow(user, newWorkflow, { projectId: 'project-1' }),
-			).rejects.toThrow('Stopping for test');
-
-			expect(userHasScopesMock).toHaveBeenCalledWith(
-				user,
-				['workflow:updateRedactionSetting'],
-				false,
-				{ projectId: 'project-1' },
-			);
-
-			const savedEntity = transactionManager.save.mock.calls[0][0] as WorkflowEntity;
-			expect(savedEntity.settings?.redactionPolicy).toBe('all');
 		});
 	});
 });
